+++ resources/lang/es/settings.php
@@ -22,6 +22,7 @@ return [
     'add_teachers' => 'Añadir profe',
     'general_preferences' => 'Preferencias generales',
     'allow_upload' => 'Permite a los y las estudiantes subir imágenes propias',
+    'disable_student_view' => 'Oculta al alumnado la información del resto',
     'show_chat' => 'Habilita el chat de la clase',
     'allow_change_class' => 'Permite cambiar la subclase',
     'allow_send_money' => 'Permite al alumnado enviarse dinero',