+++ resources/views/layouts/student.blade.php
@@ -28,15 +28,18 @@
     <a href="/classroom/show/{{ $class->code }}" class="navbar-item" style="border-right: 1px solid #aaa; border-radius: 10px;min-width: 120px" v-tippy content="{{ $student->name }}">
         <img class="rounded" src="{{ $student->avatar }}"> <span style="min-width: 70px" class="ml-2 has-text-centered">{{ Str::limit($student->name, 8, $end='...') }}</span>
     </a>
+    @php
+    settings()->setExtraColumns(['classroom_id' => $class->id]);
+    $hide = settings()->get('disable_student_view', 0);
+    @endphp
+    @if($hide != 1)
     <a href="/classroom/show/{{ $class->code }}/users" class="navbar-item">
         <i class="fad fa-users mr-2"></i> {{ __('menu.students') }}
     </a>
+    @endif
     <a href="/classroom/show/{{ $class->code }}/challenges" class="navbar-item">
         <i class="fad fa-pen-fancy mr-2"></i> {{ __('menu.challenges') }}
     </a>
-    @php
-    settings()->setExtraColumns(['classroom_id' => $class->id]);
-    @endphp
     @if(settings()->get('active_map'))
     <a href="/classroom/show/{{ $class->code }}/map" class="navbar-item">
         <i class="fad fa-map-marked-alt mr-2"></i> {{ __('menu.map') }}