+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 ## v1.67 Beta (16/08/2021)
 ### Added
-- Back button to return from viewing as a student.
 - Classroom / group roles. Now you can assign roles to students.
+- Back button to return from viewing as a student.
 
 ### Changed
 - Fixed superhero movement and positions.